refactor: move product details component to it's own folder
