+++ src/App.tsx
@@ -2,8 +2,8 @@ import { Routes, Route } from 'react-router-dom';
 import { useState } from 'react';
 import { User } from './types';
 import Home from './pages/Home/Home';
-import ProductDetails from './pages/ProductDetails';
 import Products from './pages/Products/Products';
+import ProductDetails from './pages/ProductDetails/ProductDetails';
 import Header from './components/Header/Header';
 import Footer from './components/Footer/Footer';
 import Cart from './pages/Cart';
